requirements.txt
streamlit
openai-whisper
spacy
textblob
pandas
fpdf2
audio-recorder-streamlit